add cleandesincrit parameter

+++ src/main/java/org/esupportail/esupagape/service/DossierService.java
@@ -5,6 +5,7 @@ import jakarta.persistence.EntityManager;
 import jakarta.persistence.TypedQuery;
 import jakarta.persistence.criteria.*;
 import org.apache.commons.lang3.BooleanUtils;
+import org.esupportail.esupagape.config.ApplicationProperties;
 import org.esupportail.esupagape.dtos.DocumentDto;
 import org.esupportail.esupagape.dtos.DossierIndividuClassDto;
 import org.esupportail.esupagape.dtos.DossierIndividuDto;
@@ -46,35 +47,27 @@ public class DossierService {
     private static final Logger logger = LoggerFactory.getLogger(DossierService.class);
 
     private final UtilsService utilsService;
-
     private final List<DossierInfosService> dossierInfosServices;
-
     private final DossierRepository dossierRepository;
-
     private final DocumentRepository documentRepository;
-
     private final DocumentService documentService;
-
     private final DossierAmenagementRepository dossierAmenagementRepository;
-
-    private final AmenagementRepository amenagementRepository;
-
     private final EntityManager em;
-
     private final LogService logService;
     private final IndividuRepository individuRepository;
+    private final ApplicationProperties applicationProperties;
 
     Map<String, String> codComposanteLabels = new HashMap<>();
 
 
-    public DossierService(UtilsService utilsService, List<DossierInfosService> dossierInfosServices, DossierRepository dossierRepository, DocumentRepository documentRepository, DocumentService documentService, DossierAmenagementRepository dossierAmenagementRepository, AmenagementRepository amenagementRepository, EntityManager em, LogService logService, IndividuRepository individuRepository) {
+    public DossierService(UtilsService utilsService, List<DossierInfosService> dossierInfosServices, DossierRepository dossierRepository, DocumentRepository documentRepository, DocumentService documentService, DossierAmenagementRepository dossierAmenagementRepository, EntityManager em, LogService logService, IndividuRepository individuRepository, ApplicationProperties applicationProperties) {
         this.utilsService = utilsService;
         this.documentRepository = documentRepository;
         this.documentService = documentService;
         this.dossierAmenagementRepository = dossierAmenagementRepository;
-        this.amenagementRepository = amenagementRepository;
         this.em = em;
         this.logService = logService;
+        this.applicationProperties = applicationProperties;
         Collections.reverse(dossierInfosServices);
         this.dossierInfosServices = dossierInfosServices;
         this.dossierRepository = dossierRepository;
@@ -605,7 +598,7 @@ public class DossierService {
     public boolean syncDossier(Long id) {
         Dossier dossier = dossierRepository.findById(id).orElseThrow();
         if (BooleanUtils.isTrue(dossier.getIndividu().getDesinscrit())) {
-            if(dossier.getYear().equals(utilsService.getCurrentYear()) && dossier.getStatusDossier().equals(StatusDossier.RECONDUIT)) {
+            if(applicationProperties.getCleanDesinscrits() && dossier.getYear().equals(utilsService.getCurrentYear()) && dossier.getStatusDossier().equals(StatusDossier.RECONDUIT)) {
                 logger.info("dossier " + id + " à supprimer, étudiant reconduit non présent " + dossier.getIndividu().getNumEtu());
                 deleteDossier(id);
             }
